Refactor to move tools into one and the same file.
